requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
google-generativeai==0.3.2
PyPDF2==3.0.1
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
requests==2.31.0
tenacity==8.2.3
pytest==8.0.0
pytest-asyncio==0.23.5
pyaudio==0.2.14
wave==0.0.2
numpy==1.26.4
scipy==1.12.0 